requirements.txt
tushare
xgboost
TA_lib
statsmodels
matplotlib
pandas
numpy
openpyxl
python-dotenv
